splinter ophys controllers; debug volume
